Generalized and add more thorough testing
Generalized ant colony by allowing any type of dict to be passed and the
optimal route will be returned in terms of the those node names. Added
type testing and the ability to specify the "start" node (where to begin
the traversal)

Added unit testing for new methods and functionality, broke up testing
into individual files to help speed testing (instead of having to wait
for the solver to walk through all the integration testing).